+++ docs/changelog.rst
@@ -4,6 +4,8 @@ Changelog
 3.0.0-beta.3 (unreleased)
 -------------------------
 
+- Refactored form-related methods on `ImportMixin` / `ExportMixin` (#1147)
+
 3.0.0-beta.2 (2022-04-21)
 -------------------------
 
@@ -44,13 +46,16 @@ This release makes some minor changes to the public API.  If you have overridden
 Deprecations
 ############
 
-This release adds some deprecations which will be removed in the 3.1 release.
+This release adds some deprecations which will be removed in a future release.
 
 - Add support for multiple resources in ModelAdmin. (#1223)
     - The `*Mixin.resource_class` accepting single resource has been deprecated and the new `*Mixin.resource_classes` accepting subscriptable type (list, tuple, ...) has been added.
     - Same applies to all of the `get_resource_class`, `get_import_resource_class` and `get_export_resource_class` methods.
 
 - Deprecated `exceptions.py` (#1372)
+
+- Refactored form-related methods on `ImportMixin` / `ExportMixin` (#1147)
+    - The following are deprecated: `get_import_form()`, `get_confirm_import_form()`, `get_form_kwargs()`, `get_export_form()`
 
 Enhancements
 ############